+++ src/test/java/com/jnape/palatable/lambda/functions/Fn1Test.java
@@ -4,7 +4,7 @@ import com.jnape.palatable.traitor.annotations.TestTraits;
 import com.jnape.palatable.traitor.runners.Traits;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import testsupport.EqualityAwareFn1;
+import testsupport.EquatableM;
 import testsupport.traits.ApplicativeLaws;
 import testsupport.traits.FunctorLaws;
 import testsupport.traits.MonadLaws;
@@ -13,6 +13,7 @@ import java.util.function.Function;
 
 import static com.jnape.palatable.lambda.adt.Maybe.just;
 import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
+import static com.jnape.palatable.lambda.functions.Fn1.fn1;
 import static com.jnape.palatable.lambda.functions.builtin.fn2.ReduceLeft.reduceLeft;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
@@ -21,8 +22,8 @@ import static org.junit.Assert.assertEquals;
 public class Fn1Test {
 
     @TestTraits({FunctorLaws.class, ApplicativeLaws.class, MonadLaws.class})
-    public Fn1<String, Integer> testSubject() {
-        return new EqualityAwareFn1<>("1", Integer::parseInt);
+    public EquatableM<Fn1<String, ?>, ?> testSubject() {
+        return new EquatableM<>(fn1(Integer::parseInt), f -> f.apply("1"));
     }
 
     @Test
@@ -35,9 +36,9 @@ public class Fn1Test {
     }
 
     @Test
-    public void fn1() {
+    public void staticFactoryMethods() {
         Function<String, Integer> parseInt = Integer::parseInt;
-        assertEquals((Integer) 1, Fn1.fn1(parseInt).apply("1"));
+        assertEquals((Integer) 1, fn1(parseInt).apply("1"));
     }
 
     @Test