+++ src/test/java/com/jnape/palatable/lambda/functor/builtin/LazyTest.java
@@ -1,9 +1,16 @@
 package com.jnape.palatable.lambda.functor.builtin;
 
+import com.jnape.palatable.lambda.functions.Fn1;
+import com.jnape.palatable.traitor.annotations.TestTraits;
+import com.jnape.palatable.traitor.runners.Traits;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import testsupport.EquatableM;
+import testsupport.traits.ApplicativeLaws;
+import testsupport.traits.FunctorLaws;
+import testsupport.traits.MonadLaws;
 
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Function;
 
 import static com.jnape.palatable.lambda.functions.builtin.fn3.Times.times;
 import static com.jnape.palatable.lambda.functor.builtin.Lazy.lazy;
@@ -12,7 +19,13 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static testsupport.Constants.STACK_EXPLODING_NUMBER;
 
+@RunWith(Traits.class)
 public class LazyTest {
+
+    @TestTraits({FunctorLaws.class, ApplicativeLaws.class, MonadLaws.class})
+    public EquatableM<Lazy<?>, Integer> testSubject() {
+        return new EquatableM<>(lazy(0), Lazy::value);
+    }
 
     @Test
     public void valueExtraction() {
@@ -41,12 +54,13 @@ public class LazyTest {
     @Test
     public void recursiveStackSafety() {
         assertEquals(STACK_EXPLODING_NUMBER,
-                     new Function<Lazy<Integer>, Lazy<Integer>>() {
+                     new Fn1<Lazy<Integer>, Lazy<Integer>>() {
                          @Override
-                         public Lazy<Integer> apply(Lazy<Integer> lazy) {
-                             return lazy.flatMap(x -> x < STACK_EXPLODING_NUMBER ? apply(lazy(x + 1)) : lazy(x));
+                         public Lazy<Integer> apply(Lazy<Integer> lazyX) {
+                             return lazyX.flatMap(x -> x < STACK_EXPLODING_NUMBER
+                                                       ? apply(lazy(x + 1))
+                                                       : lazy(x));
                          }
-                     }.apply(lazy(0))
-                             .value());
+                     }.apply(lazy(0)).value());
     }
 }